requirements.txt
streamlit
google-api-python-client
google-generativeai
faster-whisper
gtts
yt-dlp
ffmpeg-python
pandas
numpy